2016.12.05 添加了eeprom，防拆按键等的程序。

+++ PRO/IAR+FreeRTOS+STM32L1xx/User/bsp/inc/adc.h
@@ -1,6 +1,6 @@
 #ifndef __ADC_H__
 #define __ADC_H__
 
-void power_adc_conversion(float *power_voltage);
+void power_adc_conversion(uint16_t *power_voltage);
 
 #endif
