+++ PRO/IAR+FreeRTOS+STM32L1xx/User/bsp/inc/hal_energy.h
@@ -1,6 +1,6 @@
 #ifndef __HAL_ENERGY__
 #define __HAL_ENERGY__
 
-void hal_get_power_voltage(float *power_voltage);
+void hal_get_power_voltage(uint16_t *power_voltage);
 
 #endif
