+++ PRO/IAR+FreeRTOS+STM32L1xx/User/bsp/src/hal_energy.c
@@ -1,6 +1,6 @@
 #include "includes.h"
 
-void hal_get_power_voltage(float *power_voltage)
+void hal_get_power_voltage(uint16_t *power_voltage)
 {
     power_adc_conversion(power_voltage);
 }
